#89 [UI] 미니 사이드바 구현
